requirements.txt
fastapi[standard]>=0.110.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
sqlalchemy>=2.0.0
psycopg[binary]>=3.1.0
alembic>=1.13.0
passlib[bcrypt]>=1.7.4
bcrypt==4.0.1
PyJWT>=2.8.0
